Make all subsytem constants into one constants file
