--- src/main/java/frc/robot/constants/HardwareMap.java
@@ -1,6 +1,0 @@
-package frc.robot.constants;
-
-public class HardwareMap {
-    public static final int DROPPER_MOTOR_ID = 28;
-    public static final int DROPPER_LIDAR_ID = 27;
-}
